Dev (#27)
* feat(android): Initial family management

* feat(backend): Add models and methods for family management

* base functionality

* chore(web): fix css

* feat(web): css improvements, separate modules for different functionalities

* feat(web): Recording home page initial

* SMALL FIXES FOR DEPLOYMENT

* some more small fixes

* small fix for merge

* feat(web): Revamped shopping list page

* feat(web): add icons on shopping lists page

* feat(web): family routing, design improvements

* Angular css (#28)

* base functionality

* chore(web): fix css

* feat(web): css improvements, separate modules for different functionalities

* feat(web): Recording home page initial

* feat(web): Revamped shopping list page

* feat(web): add icons on shopping lists page

* feat(web): family routing, design improvements

* feat(web): Join requests initial

---------


* fix(web): login changes, potential fix

* fix(backend, android): List improvements

* small fix for CORS

* API security update

* validation fixes

* Web edit family (#29)

* feat(web): Family changes. Toaster

* feat(web): More family improvements. Join and leave refactor

* fix(family): Create family. JWT validation fixes

* fix(web,backend): Near finished family functionality

* feat(web):Update list name. Custom context menu

* feat(web): Search for lists

* feat(backend): Global exception handling

* feat(backend,web):bulk check

* chore(web): Fix some angular build errors (#30)

* feat(backend): bulk uncheck

* fix(web): Registration flow, logout, padding and sizing fixes

* feat(web): Recorder recording status

* fix(web): Remove shopping list item on transcript button

* fix(web): lint errors

* fix(web): Cancel recording on transcript click

* feat(web): Add shopping list items to family or create a personal list

* displaying messsage if no requests for joining family

* Pr 62 (#31)

* feat(web): invite users

* fix: add missing files

* feat(backend): Support for addedBy in shopping lists (#32)

* Pr 63 (#33)

* feat(backend): Support for addedBy in shopping lists

* feat(web): Family shopping list

* unchecking items inside shopping list

* fix

* changes after pr

* leave family/ delete family for owner

* fix(android): Update family name and code

* fix(android): Family shopping list (check, delete)

* fix(android): Family shopping list (check, delete) (#36)

* account setup

* adding settings file

* dissapearing list fix

* MERGE

* fixed (#38)

* fix (#39)

* fix(android): Different android fixes

* fix(recorder): Fix recorder funky positioning

* Android fixes (#40)

* fix(android): Family shopping list (check, delete)

* fix(android): Different android fixes

* fix(recorder): Fix recorder funky positioning

* feat(android): Account setup. Record list fixes

* fix(android): Add items by voice error handling

* router fix

* imports and remove double function

* fix join requests

* fix(backend): proper return type for slis for users

* fix(web, backend): idk

* fix(web): unfuck the fucked recorder

* fix(web): slis overflow

---------

+++ app/backend/src/main/java/com/wishlist/controllers/InvitationController.java
@@ -1,6 +1,8 @@
 package com.wishlist.controllers;
 
+import com.wishlist.exceptions.UserNotAuthorizedException;
 import com.wishlist.models.Invitation;
+import com.wishlist.security.JwtValidator;
 import com.wishlist.services.interfaces.IInvitationService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,33 +15,43 @@ import java.util.List;
 public class InvitationController {
 
     private final IInvitationService invitationService;
+    private final JwtValidator jwtValidator;
 
-    public InvitationController(IInvitationService invitationService) {
+    public InvitationController(final IInvitationService invitationService, final JwtValidator jwtValidator) {
         this.invitationService = invitationService;
+        this.jwtValidator = jwtValidator;
     }
 
-    @GetMapping
+/*    @GetMapping
     public List<Invitation> getAll() {
         return invitationService.getAllInvitations();
-    }
+    }*/
 
     @GetMapping("/{id}")
-    public ResponseEntity<Invitation> getInvitation(@PathVariable String id) {
-        Invitation invitation = invitationService.findById(id);
-        if (invitation!=null) {
-            return new ResponseEntity<>(invitation, HttpStatus.OK);
+    public ResponseEntity<Invitation> getInvitation(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
+        if (jwtValidator.validateInvitation(jwt, id)) {
+            Invitation invitation = invitationService.findById(id);
+            if (invitation != null) {
+                return new ResponseEntity<>(invitation, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            throw new UserNotAuthorizedException();
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable String id) {
-        boolean success = invitationService.delete(id);
-        if (success) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<Void> delete(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
+        if (jwtValidator.validateInvitation(jwt, id)) {
+            boolean success = invitationService.delete(id);
+            if (success) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            throw new UserNotAuthorizedException();
         }
     }
